==> Scripts/05 Analysis for OA impact asssessment.R <==
# PVGA Analysis for impact assessment

# Goal: To produce tables and charts which are part of annex to OA Policy Impact Assessment.

# NB. This only includes analysis which was used in the report - for other analysis (e.g. at a journal level) see superceded code 'PVGA Analysis - in report order'

# NB.2  see documentation for description of derived variables used in analysis below

#XXXXXXXXXXXXXXX
# Clear work space
rm(list=ls())

#library(readr)
library(tidyverse)
library(openxlsx)
library(janitor)
# library(knitr)
library(ggplot2)
# library(psych)

#XXXXXXXXX
# Import data
load("Data/merged_pvga.Rda")

####

# SET UP STYLES FOR CHARTS etc.----
#build UKRI colour palette - # UKRI brand guidelines are saved here https://www.ukri.org/files/brand-guidelines-a-guide-to-understanding-and-using-our-brand/

ukri_pal <- c("#2E2D62","#FBBB10", "#F08900","#E355EC", "#923D9D","#34D5AE", "#16978A","#FF5A5A", "#CB3564","#00BED5", "#008AAD","#67c04D", "#3E863E","#1E5DF8", "#003088")

colour_ukri <- c("#2E2D62")
grey_ukri <- c("#676767")

# colour function for box plots, bar plots etc
scale_fill_manual(values = ukri_pal)
# colour function for lines and points
scale_color_manual(values = ukri_pal)

# Set black-and-white colour scheme and add UKRI font (Arial)

theme_set(theme_bw())
theme_update(text= element_text(family="arial", size=12), 
             rect = element_blank(),
             plot.title = element_text(family="arial", hjust = 0.5),
             axis.text = element_text(family = "arial", size = 12))

# LIMITATIONS----

# Breakdown of license information ----

  # for green policies (excluding Pure Gold since we believe that these are not green policies per se)
all_green_licenses_exc_pure <- merged_pvga %>%
  filter(!is.na(g_license_single), journal_type != "Pure Gold") %>%
  count(g_license_single, .drop = FALSE) %>%
  mutate(percent = round(n / sum(n) * 100,1)) %>%
  mutate(g_license_single = factor(g_license_single, ordered = TRUE, levels = c("cc_by", "cc_by_sa", "cc_by_nd", "cc_by_nc", "cc_by_nc_sa", "cc_by_nc_nd", "bespoke_license", "all_rights_reserved", "no license requirement"))) %>%
  arrange(g_license_single) %>%
  adorn_totals("row")

  # for gold policies
all_fee_licenses <- merged_pvga %>%
  filter(!is.na(fee_license_single)) %>%
  count(fee_license_single, .drop = FALSE) %>%
  mutate(percent = round(n / sum(n) * 100,1)) %>%
  mutate(fee_license_single = factor(fee_license_single, ordered = TRUE, levels = c("cc_by", "cc_by_sa", "cc_by_nd", "cc_by_nc", "cc_by_nc_sa", "cc_by_nc_nd", "bespoke_license", "all_rights_reserved", "no license requirement"))) %>%
  arrange(fee_license_single) %>%
  adorn_totals("row")

  # bind green and gold
all_licenses <- bind_cols(all_green_licenses_exc_pure, all_fee_licenses)

write.xlsx(all_licenses, file = "Output/Tables/All licenses.xlsx")

rm(all_fee_licenses, all_green_licenses_exc_pure)

# IMPACT OF EACH POLICY SCENARIO----

# Ultimately creating a stacked bar chart comparing actual OA, potential compliance with the current policy and all three potential compliance scenarios (current, funding for all, funding for hybrid with TAs, funding for fully OA only).

  # Potential compliance with current policy
compliance_current_a <- merged_pvga %>%
  count(compliance_current, .drop = FALSE) %>%
  mutate(percent = round(n / sum(n) * 100,1), cml = round(cumsum(percent)), Scenario = "Current\nRCUK policy")

  # Potential compliance with Scenario 1
compliance_h_a <- merged_pvga %>%
  count(compliance_new_hybrid, .drop = FALSE) %>%
  mutate(percent = round(n / sum(n) * 100,1), cml = round(cumsum(percent)), Scenario = "S1 (all gold\nOA supported)")

  # Potential compliance with Scenario 2
compliance_new_a <- merged_pvga %>%
  count(compliance_new, .drop = FALSE) %>%
  mutate(percent = round(n / sum(n) * 100,1), cml = round(cumsum(percent)), Scenario = "S2a (only hybrid\ngold w/ TA)\n(current TAs)")

#  # Potential compliance with Scenario 2 (if JISC Target TAs achieved)
compliance_new_target_TAs_a <- merged_pvga %>%
  count(compliance_new_target_TAs, .drop = FALSE) %>%
  mutate(percent = round(n / sum(n) * 100,1), cml = round(cumsum(percent)), Scenario = "S2b (only hybrid\ngold w/ TA)\n(target TAs)")

  # Potential compliance with Scenario 3
compliance_new_pure_a <- merged_pvga %>%
  count(compliance_new_pure, .drop = FALSE) %>%
  mutate(percent = round(n / sum(n) * 100,1), cml = round(cumsum(percent)), Scenario = "S3 (hybrid\ngold not\nsupported")

# Bind together figures for each scenario
policy_impact_a <- bind_rows(compliance_current_a, compliance_h_a, compliance_new_a, compliance_new_target_TAs_a, compliance_new_pure_a) %>%
  unite(Compliance_Route, c('compliance_current', 'compliance_new_hybrid', 'compliance_new', 'compliance_new_target_TAs', 'compliance_new_pure'), na.rm = TRUE) %>%
  relocate(5,1,2,3,4)

  # Label each compliance route
policy_impact_a[c(1,5,9,13,17), 'Compliance_Route'] <- "Gold OA in\nfully oa journal\n"
policy_impact_a[c(2,6,10,14), 'Compliance_Route'] <- "Gold OA in\nhybrid gold journal\n(with TAs for S2)"
policy_impact_a[c(3,7,11,15,18), 'Compliance_Route'] <- "\nConfirmed green OA\n"
policy_impact_a[c(4,8,12,16,19), 'Compliance_Route'] <- "None identified under\ncurrent journal policies\n"

policy_impact_a$Compliance_Route <- factor(policy_impact_a$Compliance_Route, ordered = TRUE, levels = c(
  "Gold OA in\nfully oa journal\n", 
  "Gold OA in\nhybrid gold journal\n(with TAs for S2)", 
  "\nConfirmed green OA\n", 
  "None identified under\ncurrent journal policies\n")) # this is needed so they appear in order in table/chart

# Create stacked bar chart for articles

policy_impact_a_bar <- ggplot(policy_impact_a, aes(fill=Compliance_Route, y=percent, x=Scenario)) +
  geom_bar(position = "stack", stat = "identity") +
  scale_fill_manual(values = c("#F08900", "#FBBB10", "#16978A", "#FF5A5A"),
                    name = "Routes to Open Access\nsupported in each\npolicy scenario",
                    labels = c("Gold OA in\nfully OA journal\n", 
                               "Gold OA in hybrid\njournal (with Transitional\nAgreement for S2)",
                               "\nConfirmed green OA\n",
                               "None identified under\ncurrent journal policies\n")) +
  theme(axis.title.x = element_text(face="bold", size=14, margin = margin(t=10)),
        axis.text.x  = element_text(vjust=0.5, size=14),
        axis.text.y  = element_text(vjust=0.5, size=14),
        title =  element_text(face="bold", size=16),
        legend.text = element_text(size=16)) +
  labs(x="Policy Scenarios", y= "% of articles")

ggsave(policy_impact_a_bar, filename = "Output/Charts/Impact of each policy scenario (articles)_target_tas.png",
       width = 12.5, height = 6, dpi = 300)

# Create summary table to compare scenarios - subset percentages only

Open.Access_a <- merged_pvga %>% # adding current OA back in for table
  count(Open.Access_ukri) %>%
  mutate(percent = round(n / sum(n) * 100,1), cml = round(cumsum(percent)), Scenario = "Current\n(actual)")

policy_impact_a_table <- bind_rows(Open.Access_a, compliance_current_a, compliance_h_a, compliance_new_a, compliance_new_target_TAs_a, compliance_new_pure_a) %>%
  unite(Compliance_Route, c('Open.Access_ukri', 'compliance_current', 'compliance_new_hybrid', 'compliance_new', 'compliance_new_target_TAs', 'compliance_new_pure'), na.rm = TRUE) %>%
  relocate(5,1,2,3,4)

# Label each compliance route
policy_impact_a_table[c(1,5,9,13,17,21), 'Compliance_Route'] <- "Gold OA in\nfully oa journal\n"
policy_impact_a_table[c(2,6,10,14,18), 'Compliance_Route'] <- "Gold OA in\nhybrid gold journal\n(with TAs for S2)"
policy_impact_a_table[c(3,7,11,15,19,22), 'Compliance_Route'] <- "\nConfirmed green OA\n"
policy_impact_a_table[c(4,8,12,16,20,23), 'Compliance_Route'] <- "None identified under\ncurrent journal policies\n"

policy_impact_a$Compliance_Route <- factor(policy_impact_a$Compliance_Route, ordered = TRUE, levels = c(
  "Gold OA in\nfully oa journal\n", 
  "Gold OA in\nhybrid gold journal\n(with TAs for S2)", 
  "\nConfirmed green OA\n", 
  "None identified under\ncurrent journal policies\n")) # this is needed so they appear in order in table/chart

policy_impact_a_summary <- policy_impact_a_table %>%
  select(-c(n, cml)) %>%
  pivot_wider(names_from = Scenario, values_from = percent) %>%
  arrange(factor(Compliance_Route, levels = c("Full gold OA", "Hybrid gold\n(with TA for S2)", "Confirmed green OA", "No confirmed supported\nroute to open access")))

policy_impact_a_summary[is.na(policy_impact_a_summary)] <- 0

openxlsx::write.xlsx(as.data.frame(policy_impact_a_summary), 'Output/Tables/policy_impact_a_summary.xlsx')

# BREAKDOWN OF ARTICLES WITHOUT SUPPORTED ROUTE TO OA IN NEW POLICy----
why_unsupported <- merged_pvga %>%
  filter(compliance_new == "not compliant") %>%
  count(journal_type, fee_license_cc_by, has_ta) %>%
  mutate(percent=round(n/sum(n)*100,1))

write.xlsx(why_unsupported, file = "Output/Tables/why unsupported by new policy.xlsx")

  # same but for Target TAs in place scenario
why_unsupported_target_TAs <- merged_pvga %>%
  filter(compliance_new_target_TAs == "No supported route") %>%
  count(journal_type, fee_license_cc_by, has_ta) %>%
  mutate(percent=round(n/sum(n)*100,2))

# ACTUAL OPEN ACCESS STATUS----
  # this is using Dimensions variable Open.Access (simplified to Open.Access_ukri which merges categories to match main UKRI categories)
  # there is a more complex version of this table showing average number of articles per OA type but not used in current version of finding so only including simple version here.

Open.Access_a <- merged_pvga %>% # article level
  count(Open.Access_ukri) %>%
  mutate(percent = round(n / sum(n) * 100,1)) %>%
  adorn_totals("row")

openxlsx::write.xlsx(as.data.frame(Open.Access_a), 'Output/Tables/Open.Access.xlsx')

# Proportion of all articles available green OA
Repository_versions <- merged_pvga %>%
  rowwise() %>%
  mutate(has_green = !is.na(upw_green_version)) %>%
  ungroup() %>%
  group_by(Open.Access_ukri, has_green) %>%
  count() %>%
  group_by(Open.Access_ukri) %>%
  mutate(percent = n/sum(n)*100)

# Licenses for green OA articles from Unpaywall----

  # needs a bit of cleaning first
merged_pvga$upw_green_licence[merged_pvga$upw_green_licence %in% c("pd", "No green license found")] <- NA
merged_pvga$upw_green_licence[merged_pvga$upw_green_licence %in% c("cc0")] <- "cc-by"
merged_pvga$upw_green_licence[merged_pvga$upw_green_licence %in% c("elsevier-specific: oa user license", "acs-specific: authorchoice/editors choice usage agreement")] <- "implied-oa"

all_green_licences_upw <- merged_pvga %>%
  filter(!is.na(upw_green_version)) %>%
  count(upw_green_licence) %>%
  mutate(percent_all_inc_w_gold = n/sum(n)*100) %>%
  rename(n_all_inc_w_gold = n)

green_licences_green_only_upw <- merged_pvga %>%
  filter(Open.Access_ukri == "Green") %>%
  count(upw_green_licence) %>%
  mutate(percent_only_green_route = n/sum(n)*100) %>%
  rename(n_only_green_route = n)

green_licences <- bind_cols(all_green_licences_upw, green_licences_green_only_upw) %>%
  mutate(upw_green_license = factor(upw_green_licence...1, levels=c("cc-by", "cc-by-sa", "cc-by-nd", "cc-by-nc", "cc-by-nc-sa", "cc-by-nc-nd", "implied-oa"), ordered = TRUE)) %>%
  select(-upw_green_licence...4) %>%
  arrange(upw_green_license)

openxlsx::write.xlsx(as.data.frame(green_licences), 'Output/Tables/green_licenses_upw.xlsx')

# Licences for gold OA from unpaywall----
merged_pvga$upw_gold_licence[merged_pvga$upw_gold_licence %in% c("acs-specific: authorchoice/editors choice usage agreement",
"elsevier-specific: oa user license", "implied-oa", "publisher-specific license", "pd")] <- "Other OA licence"

<<<<<<< HEAD
gold_licences_upw_all <- merged_pvga %>%
=======
gold_licences_upw <- merged_pvga %>%
>>>>>>> 32b4db98d7cc7a060abdcae2fbdb7449c7c31d08
  filter(Open.Access_ukri %in% c("Pure Gold", "Hybrid gold")) %>%
  mutate(upw_gold_licence = factor(upw_gold_licence, levels = c("cc-by", "cc-by-sa", "cc-by-nd", "cc-by-nc", "cc-by-nc-sa", "cc-by-nc-nd", "Other OA licence"),
         ordered = TRUE)) %>%
  count(upw_gold_licence) %>%
  mutate(percent = n/sum(n)*100)

<<<<<<< HEAD
gold_licences_upw_journal_type <- merged_pvga %>%
  filter(Open.Access_ukri %in% c("Pure Gold", "Hybrid gold")) %>%
  mutate(upw_gold_licence = factor(upw_gold_licence, levels = c("cc-by", "cc-by-sa", "cc-by-nd", "cc-by-nc", "cc-by-nc-sa", "cc-by-nc-nd", "Other OA licence"),
                                   ordered = TRUE)) %>%
  group_by(journal_type) %>%
  count(upw_gold_licence) %>%
  mutate(percent = n/sum(n)*100) %>%
  pivot_wider(names_from = journal_type, values_from = (c(n, percent)))

gold_licenses_all_and_journal_type <- gold_licences_upw_all %>%
  left_join(gold_licences_upw_journal_type, by = "upw_gold_licence") %>%
  relocate(upw_gold_licence, n, percent, `n_Pure Gold`, `percent_Pure Gold`, n_Hybrid, percent_Hybrid)

openxlsx::write.xlsx(as.data.frame(gold_licenses_all_and_journal_type), 'Output/Tables/gold_licenses_upw.xlsx')
=======
openxlsx::write.xlsx(as.data.frame(gold_licences_upw), 'Output/Tables/gold_licenses_upw.xlsx')
>>>>>>> 32b4db98d7cc7a060abdcae2fbdb7449c7c31d08


# Actual Open Access x Discipline ----

disciplines <- merged_pvga %>% 
  separate(discipline, into = c("disc1", "disc2", "disc3", "disc4"), sep = "; ") %>%
  pivot_longer(c(disc1, disc2, disc3, disc4), names_to = "disc_all", values_to ="disc")

Open.Access_a_disc <- disciplines %>%
  filter(!is.na(disc)) %>%
  group_by(disc) %>%
  count(Open.Access_ukri, .drop = FALSE) %>%
  mutate(percent = (n / sum(n)) * 100) %>%
  mutate(cml = round(cumsum(percent),1)) %>%
  mutate(percent = round(percent, 1)) %>%
  mutate(disc = factor(disc, ordered = TRUE, levels = c("Arts & Humanities", "Health Sciences", "Life Sciences", "Physical Sciences", "Social Sciences", "Missing")), 
         Open.Access_ukri = factor(Open.Access_ukri, ordered = TRUE, levels = c("Pure Gold", "Hybrid gold", "Green", "Closed"))) %>% # this is needed to get everything in correct order
  arrange(disc, Open.Access_ukri)

# Bar
Open.Access_a_disc_bar <- Open.Access_a_disc %>%
  ggplot(aes(x=disc, y=percent)) +
  geom_bar(aes(fill=Open.Access_ukri), stat = "identity", position = "stack", width = 0.9) +
  scale_fill_manual(values = c("#F08900", "#FBBB10", "#16978A", "#FF5A5A"), name = "Open Access Category", labels = c("Gold OA in fully\nOA journal\n", "Gold OA in Hybrid\njournal\n", "Green OA (Version\nof Record or Author\nAccepted Manuscript", "\nNo OA version identified\n")) +
  # ggtitle("Open Access Categories by discipline") +
  theme(axis.title.x = element_text(face="bold", size=16, margin = margin(t=10)),
        axis.text.x  = element_text(vjust=0.5, size=14),
        axis.title.y = element_text(face="bold", size=16, margin = margin(t=10)),
        axis.text.y  = element_text(vjust=0.5, size=14),
        title =  element_text(face="bold", size=16),
        legend.text = element_text(size=14)) +
  labs(x="Disciplines", y= "% of all articles in panel", fill = "Open access") +
  scale_x_discrete(labels = c("Arts &\nHumanities", "Health\nSciences","Life\nSciences", "Physical\nSciences", "Social\nSciences", "Missing"))

ggsave(Open.Access_a_disc_bar, filename = "Output/Charts/Open.Access_a_disc_bar.png", width = 9, height = 5, dpi = 300)

# Repository versions by discipline (including those with gold OA version)----
disciplines <- merged_pvga %>% 
  separate(discipline, into = c("disc1", "disc2", "disc3", "disc4"), sep = "; ") %>%
  pivot_longer(c(disc1, disc2, disc3, disc4), names_to = "disc_all", values_to ="disc")

Repo_versions_a_disc <- disciplines %>%
  filter(!is.na(disc)) %>%
  group_by(disc, upw_green_version) %>%
  count() %>%
  mutate(disc = factor(disc, ordered = TRUE, levels = c("Arts & Humanities", "Health Sciences", "Life Sciences", "Physical Sciences", "Social Sciences", "Missing"))) %>%
  group_by(disc) %>%
  mutate(percent = n/sum(n)*100) %>%
  arrange(disc) %>%
  pivot_wider(names_from = upw_green_version, values_from = c(n, percent)) %>%
  rowwise() %>%
  mutate(all_repo_versions_n = n_acceptedVersion + n_publishedVersion, all_repo_versions_percent = percent_acceptedVersion + percent_publishedVersion) %>%
  select(disc, all_repo_versions_n, all_repo_versions_percent)

# Merging current OA status tables for report----
Open.Access_and_repositories <- Open.Access_a_disc %>%
  select(-cml, -n) %>%
  pivot_wider(names_from = Open.Access_ukri, values_from = percent) %>%
  left_join(Repo_versions_a_disc, by = "disc") %>%
  left_join(disciplines %>%
              filter(!is.na(disc)) %>%
              mutate(disc = factor(disc, ordered = TRUE, levels = c("Arts & Humanities", "Health Sciences", "Life Sciences", "Physical Sciences", "Social Sciences", "Missing"))) %>%
              count(disc) %>% rename(n_all_disc = n)) %>%
  select(disc, n_all_disc, `Pure Gold`, `Hybrid gold`, Green, Closed, all_repo_versions_percent)

openxlsx::write.xlsx(as.data.frame(Open.Access_and_repositories), 'Output/Tables/Open.Access_and_repositories.xlsx')


#  Actual open access x publisher----
top_publishers_oac <- merged_pvga %>%
  group_by(Publisher) %>%
  count(Open.Access_ukri) %>%
  pivot_wider(names_from = Open.Access_ukri, values_from = n) %>%
  adorn_totals("col") %>%
  rename(pure_gold = "Pure Gold", hybrid_gold = "Hybrid gold") %>%
  mutate(pure_percent = round(pure_gold/Total*100,1), hybrid_gold_percent = round(hybrid_gold/Total*100,1), green_percent = round(Green/Total*100,1), closed_percent = round(Closed/Total*100,1)) %>%
  select(-c(2,3,4,5)) %>%
  rename(Total_n_articles = Total) %>%
  mutate(percent_of_total_articles = round(Total_n_articles/sum(Total_n_articles)*100,1)) %>%
  arrange(desc(Total_n_articles)) %>%
  mutate(cml = round(cumsum(percent_of_total_articles),0)) %>%
  relocate(1,2,7,8,3,4,5,6) %>%
  adorn_totals("row")

openxlsx::write.xlsx(as.data.frame(top_publishers_oac), 'Output/Tables/top_publishers_oac.xlsx')


# Actual open access x research organisation----
universities <- merged_pvga %>% 
  separate(Research.Organizations...standardized, into = c("uni1", "uni2", "uni3", "uni4", "uni5", "uni6", "uni7", "uni8","uni9", "uni10",
                                                           "uni11", "uni12", "uni13", "uni14", "uni15", "uni16"), sep = "; ") %>%
  pivot_longer(c(uni1, uni2, uni3, uni4, uni5, uni6, uni7, uni8, uni9, uni10, uni11, uni12, uni13, uni14, uni15, uni16),
               names_to = "uni_all", values_to ="uni")


Open.Access_a_uni <- universities %>%
  filter(!is.na(uni)) %>%
  mutate(Open.Access_ukri = factor(Open.Access_ukri, ordered = TRUE, levels = c("Pure Gold", "Hybrid gold", "Green", "Closed"))) %>% # this is needed to get everything in correct order
  group_by(uni) %>%
  count(Open.Access_ukri, .drop = FALSE) %>%
  mutate(percent = (n / sum(n)) * 100) %>%
  mutate(cml = round(cumsum(percent),1)) %>%
  mutate(percent = round(percent, 1)) %>%
  group_by(uni) %>%
  summarise(group_size = sum(n),
            gold = sum(percent[Open.Access_ukri %in% c("Pure Gold", "Hybrid gold")]),
            green = percent[Open.Access_ukri == "Green"],
            compliant = sum(percent[Open.Access_ukri %in% c("Pure Gold", "Hybrid gold", "Green")])) %>%
  arrange(desc(group_size), uni) %>%
  mutate(proportion_all_articles = group_size/sum(group_size)*100) %>%
  filter(group_size > 39)

openxlsx::write.xlsx(as.data.frame(Open.Access_a_uni), 'Output/Tables/Open.Access_a_uni.xlsx')



# VARIATION OF POLICY IMPACT BY GROUP----

# 1. Variation by UKRI funder----
ukri_funders <- merged_pvga %>% 
  separate(ukri_funders, into = c("ukri_funder1", "ukri_funder2", "ukri_funder3", "ukri_funder4","ukri_funder5", "ukri_funder6", "ukri_funder7", "ukri_funder8"), sep = ", ") %>%
  pivot_longer(c(ukri_funder1, ukri_funder2, ukri_funder3, ukri_funder4, ukri_funder5, ukri_funder6, ukri_funder7, ukri_funder8), names_to = "ukri_funder_all", values_to ="ukri_funder")

compliance_new_ukri_funder <- ukri_funders  %>%
  filter(!is.na(ukri_funder)) %>% 
  group_by(ukri_funder) %>%
  count(compliance_new) %>%
  mutate(percent = round(n/sum(n)*100,1)) %>%
  mutate(total_n = sum(n)) %>%
  select(-n) %>%
  pivot_wider(names_from = compliance_new, values_from = c(percent, total_n)) %>%
  mutate(total_supported = sum(across(1:3), na.rm = TRUE)) %>% # I couldn't get rowsum to work - for some weird reason this code seems to requires 0 indexing of columns (i.e. I'm adding columns 2:4 in this line)
  rename(total_n = "total_n_c: pure gold") %>%
  select(-c(`total_n_c: hybrid gold with a TA`, `total_n_c: confirmed green oa`, `total_n_not compliant`)) %>%
  relocate(`ukri_funder`,total_supported)

# Add column to show proportion of each division covered by TA
TA_ukri_funder <- ukri_funders %>%
  filter(!is.na(ukri_funder), journal_type == "Hybrid") %>%
  group_by(ukri_funder) %>%
  count(has_ta, .drop = FALSE) %>%
  mutate(percent = round(n/sum(n)*100,1)) %>%
  filter(has_ta == "yes")

# Merge in with compliance_new_ukri_funder to add TA column
compliance_new_ukri_funder_ta <- bind_cols(compliance_new_ukri_funder, TA_ukri_funder) %>%
  select(-c(`ukri_funder...8`,has_ta, n)) %>%
  rename(percent_hybrid_with_TA = percent)

# Add all articles row (this feels very elaborate but I couldn't think of a simpler way to do it)
compliance_all_articles <- as.data.frame(t(merged_pvga %>%
                                             count(compliance_new) %>%
                                             mutate(percent = round(n/sum(n)*100,1)))) %>%
  mutate(total_n = nrow(merged_pvga)) %>%
  slice(3)

compliance_all_articles <- as.data.frame(lapply(compliance_all_articles,as.numeric))
compliance_all_articles <- compliance_all_articles %>%
  mutate(total_supported = sum(across(1:3))) %>%
  relocate(total_supported)

ta_all_hybrid <- merged_pvga %>% filter(journal_type == "Hybrid") %>% count(has_ta) %>% mutate(percent_hybrid_with_TA = round(n/sum(n)*100,1)) %>% filter(has_ta == "yes")

compliance_ta_all_articles <- bind_cols(compliance_all_articles, ta_all_hybrid) %>%
  select(-c(has_ta, n))

colnames(compliance_ta_all_articles) <- colnames(compliance_new_ukri_funder_ta[2:8])

# Merge in with compliance_new_ukri_funder_ta to create final table
compliance_new_ukri_funder_ta <- bind_rows(compliance_new_ukri_funder_ta, compliance_ta_all_articles)
compliance_new_ukri_funder_ta[12,'ukri_funder'] <- "All articles"

openxlsx::write.xlsx(as.data.frame(compliance_new_ukri_funder_ta), 'Output/Tables/compliance_new_ukri_funder_ta.xlsx')
rm(compliance_new_ukri_funder, compliance_all_articles, ta_all_hybrid, TA_ukri_funder)


# Work it out with TA targets in place
compliance_new_ukri_funder_target_TAs <- ukri_funders  %>%
  filter(!is.na(ukri_funder)) %>% 
  group_by(ukri_funder) %>%
  count(compliance_new_target_TAs) %>%
  mutate(percent = round(n/sum(n)*100,1)) %>%
  mutate(total_n = sum(n))

rm(ukri_funders)


# 2. Variation by discipline ----

disciplines <- merged_pvga %>% 
  separate(discipline, into = c("disc1", "disc2", "disc3", "disc4"), sep = "; ") %>%
  pivot_longer(c(disc1, disc2, disc3, disc4), names_to = "disc_all", values_to ="disc")

compliance_new_disc <- disciplines %>%
  filter(!is.na(disc)) %>%
  group_by(disc) %>%
  count(compliance_new) %>%
  mutate(percent = round(n/sum(n)*100,1)) %>%
  mutate(total_n = sum(n)) %>%
  select(-n) %>%
  pivot_wider(names_from = compliance_new, values_from = c(percent, total_n)) %>%
  mutate(total_supported = sum(across(1:3), na.rm = TRUE)) %>% # I couldn't get rowsum to work - for some weird reason this code seems to requires 0 indexing of columns (i.e. I'm adding columns 2:4 in this line)
  rename(total_n = "total_n_c: pure gold") %>%
  select(-c(7,8,9)) %>%
  relocate(1,6)

compliance_new_disc$disc[is.na(compliance_new_disc$disc)] <- "Missing"

    # Add column to show proportion of each division covered by TA
TA_disc <- disciplines %>%
  filter(!is.na(disc), journal_type == "Hybrid") %>%
  group_by(disc) %>%
  count(has_ta, .drop = FALSE) %>%
  mutate(percent = round(n/sum(n)*100,1)) %>%
  filter(has_ta == "yes")

    # Merge in with compliance_new_disc to add TA column
compliance_new_disc_ta <- left_join(compliance_new_disc, TA_disc, by = "disc") %>%
  select(-c(has_ta, n)) %>%
  rename(percent_hybrid_with_TA = percent) %>%
  relocate(disc, total_supported, `percent_c: pure gold`,	`percent_c: hybrid gold with a TA`,	`percent_c: confirmed green oa`,	`percent_not compliant`
           , total_n, percent_hybrid_with_TA)

openxlsx::write.xlsx(as.data.frame(compliance_new_disc_ta), 'Output/Tables/compliance_new_disc_ta.xlsx')

# what does it look like with target TAs in place?
compliance_new_disc_target_TAs <- disciplines %>%
  filter(!is.na(disc)) %>%
  group_by(disc) %>%
  count(compliance_new_target_TAs) %>%
  mutate(percent = round(n/sum(n)*100,1)) %>%
  mutate(total_n = sum(n)) %>%
  select(-n) %>%
  pivot_wider(names_from = compliance_new_target_TAs, values_from = c(percent, total_n)) %>%
  mutate(total_supported = sum(across(1:3), na.rm = TRUE)) %>% # I couldn't get rowsum to work - for some weird reason this code seems to requires 0 indexing of columns (i.e. I'm adding columns 2:4 in this line)
  rename(total_n = "total_n_c: pure gold") %>%
  select(-c(7,8,9,10)) %>%
  relocate(1,6)

compliance_new_disc$disc[is.na(compliance_new_disc$disc)] <- "Missing"

    # Stacked bar chart for Compliance by discipline
compliance_new_disc_a <- disciplines %>%
  filter(!is.na(disc)) %>% 
  group_by(disc) %>%
  count(compliance_new, .drop = FALSE) %>%
  mutate(percent = round(n/sum(n)*100,1))

compliance_new_disc_a$disc[1:4] <- "Arts &\nHumanities"
compliance_new_disc_a$disc[5:8] <- "Health\nSciences"
compliance_new_disc_a$disc[9:12] <- "Life\nSciences"
compliance_new_disc_a$disc[13:16] <- "Missing"
compliance_new_disc_a$disc[17:20] <- "Physical\nSciences"
compliance_new_disc_a$disc[21:24] <- "Social\nSciences"

compliance_new_disc_a$disc <- factor(compliance_new_disc_a$disc, ordered = TRUE, c("Arts &\nHumanities", "Health\nSciences", "Life\nSciences", "Physical\nSciences", "Social\nSciences", "Missing"))

compliance_new_disc_chart <- ggplot(compliance_new_disc_a, aes(fill=compliance_new, y = percent, x = disc)) +
  geom_bar(position = "stack", stat = "identity", width = 0.9) +
  scale_fill_manual(values = c("#F08900", "#FBBB10", "#16978A", "#FF5A5A"), name = "Supported routes to\nOpen Access", labels = c("Gold OA in\nfully OA journal\n", "Gold OA in Hybrid\njournal with\nTransitional Agreement", "\nConfirmed green OA\n", "None identified under\ncurrent journal policies\n")) +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  # scale_fill_discrete(name = "Supported routes to\nOpen Access", labels = c("Gold OA in\nfully OA journal\n", "Gold OA in Hybrid\n journal with TA", "Confirmed green OA", "None identified under\ncurrent journal policies")) +
  theme(axis.title.x = element_text(face="bold", size=14, margin = margin(t=10)),
        axis.text.x  = element_text(vjust=0.5, size=12),
        title =  element_text(face="bold", size=16),
        legend.text = element_text(size=14)) +
  labs(x="Discipline", y= "% of articles")

ggsave(compliance_new_disc_chart, filename = "Output/Charts/compliance_new_disc_chart.png", width = 8, height = 5, dpi = 300)



# 3. Variation by Fields of Research Division (i.e. subject)----

# FIRST DIVISIONS
divisions <- merged_pvga %>% 
  separate(for_division, into = c("division1", "division2", "division3", "division4"), sep = "; ") %>%
  pivot_longer(c(division1, division2, division3, division4), names_to = "division_all", values_to ="division")

compliance_new_division <- divisions %>%
  filter(!is.na(division)) %>%
  group_by(division) %>%
  count(compliance_new) %>%
  mutate(percent = round(n/sum(n)*100,1)) %>%
  mutate(total_n = sum(n)) %>%
  select(-n) %>%
  pivot_wider(names_from = compliance_new, values_from = c(percent, total_n)) %>%
  mutate(total_supported = sum(across(1:3), na.rm = TRUE)) %>% # I couldn't get rowsum to work - for some weird reason this code seems to requires 0 indexing of columns (i.e. I'm adding columns 2:4 in this line)
  unite(total_n, c(`total_n_c: pure gold`, `total_n_c: hybrid gold with a TA`, `total_n_c: confirmed green oa`, `total_n_not compliant`), na.rm = TRUE) %>%
  mutate(total_n = gsub(".*_", "", total_n)) %>%
  relocate(division,total_n)

compliance_new_division$division[is.na(compliance_new_division$division)] <- "Missing"


      # Add column to show proportion of each division covered by TA

TA_div <- divisions %>%
  filter(!is.na(division), journal_type == "Hybrid") %>%
  group_by(division) %>%
  count(has_ta, .drop = FALSE) %>%
  mutate(percent = round(n/sum(n)*100,1)) %>%
  filter(has_ta == "yes")

  # Merge in with compliance_new_division
compliance_new_division_ta <- bind_cols(compliance_new_division, TA_div) %>%
  select(-c(division...8,has_ta, n)) %>%
  rename(percent_hybrid_with_TA = percent) %>%
  relocate(division...1, total_supported, `percent_c: pure gold`,	`percent_c: hybrid gold with a TA`,	`percent_c: confirmed green oa`,	`percent_not compliant`
           , total_n, percent_hybrid_with_TA)

openxlsx::write.xlsx(as.data.frame(compliance_new_division_ta), 'Output/Tables/compliance_new_division_ta.xlsx')


#what does it look like with target TAs?

compliance_new_division_target_TAs <- divisions %>%
  filter(!is.na(division)) %>%
  group_by(division) %>%
  count(compliance_new_target_TAs) %>%
  mutate(percent = round(n/sum(n)*100,2)) %>%
  mutate(total_n = sum(n)) %>%
  select(-n) %>%
  pivot_wider(names_from = compliance_new_target_TAs, values_from = c(percent, total_n)) %>%
  mutate(total_supported = sum(across(1:3), na.rm = TRUE)) %>% # I couldn't get rowsum to work - for some weird reason this code seems to requires 0 indexing of columns (i.e. I'm adding columns 2:4 in this line)
  unite(total_n, c(`total_n_c: pure gold`, `total_n_c: hybrid gold with a TA`, `total_n_c: confirmed green oa`, `total_n_No supported route`), na.rm = TRUE) %>%
  mutate(total_n = gsub(".*_", "", total_n)) %>%
  relocate(division,total_n)

compliance_new_division_target_TAs$division[is.na(compliance_new_division_target_TAs$division)] <- "Missing"

# 3. Variation by Fields of Research Group (i.e. specialism)----
groups <- merged_pvga %>% 
  separate(for_group, into = c("group1", "group2", "group3", "group4"), sep = "; ") %>%
  pivot_longer(c(group1, group2, group3, group4), names_to = "group_all", values_to ="group")

compliance_new_group <- groups %>%
  filter(!is.na(group)) %>%
  group_by(group) %>%
  count(compliance_new) %>%
  mutate(percent = round(n/sum(n)*100,1)) %>%
  mutate(total_n = sum(n)) %>%
  select(-n) %>%
  pivot_wider(names_from = compliance_new, values_from = c(percent, total_n)) %>%
  mutate(total_supported = sum(across(1:3), na.rm = TRUE)) %>% # I couldn't get rowsum to work - for some weird reason this code seems to requires 0 indexing of columns (i.e. I'm adding columns 2:4 in this line)
  unite(total_n, c(`total_n_c: pure gold`, `total_n_c: hybrid gold with a TA`, `total_n_c: confirmed green oa`, `total_n_not compliant`), na.rm = TRUE) %>%
  mutate(total_n = gsub(".*_", "", total_n)) %>%
  relocate(group,total_n)

compliance_new_group$group[is.na(compliance_new_group$group)] <- "Missing"

    # Proportion of specialisms covered by TAs
TA_group <- groups %>%
  mutate(has_ta = factor(has_ta)) %>% # this is needed to stop specialisms with 0% TA coverage being dropped
  filter(!is.na(group), journal_type == "Hybrid") %>%
  group_by(group) %>%
  count(has_ta, .drop = FALSE) %>%
  mutate(percent = round(n/sum(n)*100,1)) %>%
  filter(has_ta == "yes")

# Merge in with compliance_new_group
compliance_new_group_ta <- bind_cols(compliance_new_group, TA_group) %>%
  select(-c(group...8, has_ta, n)) %>%
  rename(percent_hybrid_with_TA =percent) %>%
  relocate(group...1, total_supported, `percent_c: pure gold`,	`percent_c: hybrid gold with a TA`,	`percent_c: confirmed green oa`,	`percent_not compliant`
           , total_n, percent_hybrid_with_TA)

openxlsx::write.xlsx(as.data.frame(compliance_new_group_ta), 'Output/Tables/compliance_new_group_ta.xlsx')


# what does it look like with Jisc target TAs in place?
compliance_new_group_target_TAs <- groups %>%
  filter(!is.na(group)) %>%
  group_by(group) %>%
  count(compliance_new_target_TAs) %>%
  mutate(percent = round(n/sum(n)*100,1)) %>%
  mutate(total_n = sum(n, na.rm = TRUE)) %>%
  select(-n) %>%
  pivot_wider(names_from = compliance_new_target_TAs, values_from = c(percent, total_n)) %>%
  mutate(total_supported = sum(across(1:3), na.rm = TRUE)) %>% # I couldn't get rowsum to work - for some weird reason this code seems to requires 0 indexing of columns (i.e. I'm adding columns 2:4 in this line)
  unite(total_n, c(`total_n_c: pure gold`, `total_n_c: hybrid gold with a TA`, `total_n_c: confirmed green oa`, `total_n_No supported route`), na.rm = TRUE) %>%
  mutate(total_n = gsub(".*_", "", total_n)) %>%
  relocate(group,total_n)

compliance_new_group_target_TAs$group[is.na(compliance_new_group_target_TAs$group)] <- "Missing"

# 4. Variation by Publisher----


# Non-compliant by publisher (current TAs)

not_compliance_new_publisher <- merged_pvga %>%
  filter(compliance_new2 == "not compliant") %>%
  count(Publisher, sort = TRUE) %>%
  mutate(percent = (n / sum(n)) * 100) %>%
  mutate(cml = cumsum(percent)) %>%
  adorn_totals("row")
# add in total articles per publisher
articles_per_publisher <- merged_pvga %>%
  count(Publisher, sort = TRUE)

not_compliance_new_publisher <- left_join(not_compliance_new_publisher, articles_per_publisher, by = "Publisher") %>%
  mutate(percent_of_publisher_total = round(n.x / n.y * 100,1))
openxlsx::write.xlsx(as.data.frame(not_compliance_new_publisher), 'Output/Tables/not_compliant_new_publisher.xlsx')


# # Non-compliant by publisher (JISC TAs in place)
not_compliance_new_publisher_target_TAs <- merged_pvga %>%
  filter(compliance_new_target_TAs == "No supported route") %>%
  count(Publisher, sort = TRUE) %>%
  mutate(percent = (n / sum(n)) * 100) %>%
  mutate(cml = cumsum(percent))

  # add in total articles per publisher
articles_per_publisher <- merged_pvga %>%
  count(Publisher, sort = TRUE)

not_compliance_new_publisher_target_TAs <- left_join(not_compliance_new_publisher_target_TAs, articles_per_publisher, by = "Publisher") %>%
  mutate(percent_of_publisher_total = round(n.x / n.y * 100,1))

openxlsx::write.xlsx(as.data.frame(not_compliance_new_publisher_target_TAs), 'Output/Tables/not_compliant_new_publisher_target_TAs.xlsx')

# Publisher by discipline
disciplines <- merged_pvga %>% 
  separate(discipline, into = c("disc1", "disc2", "disc3", "disc4"), sep = "; ") %>%
  pivot_longer(c(disc1, disc2, disc3, disc4), names_to = "disc_all", values_to ="disc")

publisher_discipline <- disciplines %>%
  filter(!is.na(disc)) %>%
  group_by(disc) %>%
  count(Publisher) %>%
  arrange(desc(n)) %>%
  arrange(disc) %>%
  mutate(percent = round(n/sum(n)*100,1), cml = cumsum(percent)) %>%
  adorn_totals("row")
  filter(cml <= 80)

openxlsx::write.xlsx(as.data.frame(publisher_discipline), 'Output/Tables/publisher_discipline.xlsx')

# Publisher by discipline long tail
  # this suggests that AHSS doesn't have a particularly long tail
publishers_per_discipline <- disciplines %>%
  filter(!is.na(disc)) %>%
  group_by(disc) %>%
  count(Publisher) %>%
  count(disc)

# Number of articles missing publisher information in each discipline
  # this suggests that AHSS doesn't have lots of articles missing publisher info
missing_publisher_info <- bind_rows(merged_pvga %>% filter(grepl("Arts", discipline)) %>% count(Publisher) %>% mutate(percent = n/sum(n)*100, discipline = "Arts") %>% filter(is.na(Publisher)),
                                    merged_pvga %>% filter(grepl("Health", discipline)) %>% count(Publisher) %>% mutate(percent = n/sum(n)*100, discipline = "Health") %>% filter(is.na(Publisher)),
                                    merged_pvga %>% filter(grepl("Life", discipline)) %>% count(Publisher) %>% mutate(percent = n/sum(n)*100, discipline = "Life") %>% filter(is.na(Publisher)),
                                    merged_pvga %>% filter(grepl("Physical", discipline)) %>% count(Publisher) %>% mutate(percent = n/sum(n)*100, discipline = "Physical") %>% filter(is.na(Publisher)),
                                    merged_pvga %>% filter(grepl("Social", discipline)) %>% count(Publisher) %>% mutate(percent = n/sum(n)*100, discipline = "Social") %>% filter(is.na(Publisher))) %>%
  relocate(discipline) %>%
  select(-Publisher)

# CURRENT SUPPORT FOR GREEN OA IN JOURNAL POLICIES----

# Simple breakdown of green compliance
simple_green <- merged_pvga %>% 
  group_by(g_license_cc_by, g_embargo_zero) %>% 
  count() %>%
  ungroup() %>%
  mutate(percent = round(n / sum(n) * 100,1))

openxlsx::write.xlsx(as.data.frame(simple_green), 'Output/Tables/simple_green_breakdown.xlsx')


# TRANSFORMATIVE AGREEMENTS----

# 1. Impact of new TAs on potential compliance with UKRI policy----

# Proportion of articles in hybrid journals covered by TAs
ta_coverage <- merged_pvga %>% filter(journal_type == "Hybrid") %>% count(ta_split) %>% mutate(percent=n/sum(n)*100)

# Create stacked bar chart comparing potential compliance with several different assumptions about coverage of Transformative Agreements

  # No Transformative Agreement option/ none negotiated
merged_pvga$compliance_new_ta_noTA <- NA
merged_pvga$compliance_new_ta_noTA[merged_pvga$num_fee %in% c(1,4)] <- "c: pure gold"
merged_pvga$compliance_new_ta_noTA[merged_pvga$ta_split == "noTA" & merged_pvga$num_fee %in% c(2,5)] <- "c: hybrid gold with a TA"
merged_pvga$compliance_new_ta_noTA[!merged_pvga$num_fee %in% c(1,4) & merged_pvga$num_new_green %in% c(1,3)] <- "c: confirmed green oa"
merged_pvga$compliance_new_ta_noTA[is.na(merged_pvga$compliance_new_ta_noTA)] <- "No supported route"

merged_pvga$compliance_new_ta_noTA <- factor(merged_pvga$compliance_new_ta_noTA, ordered = TRUE, levels = c("c: pure gold", "c: hybrid gold with a TA", "c: confirmed green oa", "No supported route"))

compliance_new_ta_noTA_a <- merged_pvga %>%
  count(compliance_new_ta_noTA, .drop = FALSE) %>%
  mutate(percent = round(n / sum(n) * 100,1), cml = round(cumsum(percent)))

  # Transformative agreements to end of 2019
merged_pvga$compliance_new_ta_2019 <- NA
merged_pvga$compliance_new_ta_2019[merged_pvga$num_fee %in% c(1,4)] <- "c: pure gold"
merged_pvga$compliance_new_ta_2019[merged_pvga$ta_split == "2019" & merged_pvga$num_fee %in% c(2,5)] <- "c: hybrid gold with a TA"
merged_pvga$compliance_new_ta_2019[!merged_pvga$num_fee %in% c(1,4) & !(merged_pvga$ta_split == "2019" & merged_pvga$num_fee %in% c(2,5)) & merged_pvga$num_new_green %in% c(1,3)] <- "c: confirmed green oa"
merged_pvga$compliance_new_ta_2019[is.na(merged_pvga$compliance_new_ta_2019)] <- "No supported route"

merged_pvga$compliance_new_ta_2019 <- factor(merged_pvga$compliance_new_ta_2019, ordered = TRUE, levels = c("c: pure gold", "c: hybrid gold with a TA", "c: confirmed green oa", "No supported route"))

compliance_new_ta_2019_a <- merged_pvga %>%
  count(compliance_new_ta_2019, .drop = FALSE) %>%
  mutate(percent = round(n / sum(n) * 100,1), cml = round(cumsum(percent)))


  # Transformative agreements to end of 2020
merged_pvga$compliance_new_ta_2020 <- NA
merged_pvga$compliance_new_ta_2020[merged_pvga$num_fee %in% c(1,4)] <- "c: pure gold"
merged_pvga$compliance_new_ta_2020[merged_pvga$ta_split %in% c("2019","2020") & merged_pvga$num_fee %in% c(2,5)] <- "c: hybrid gold with a TA"
merged_pvga$compliance_new_ta_2020[!merged_pvga$num_fee %in% c(1,4) & !((merged_pvga$ta_split == "2019" | merged_pvga$ta_split == "2020") & merged_pvga$num_fee %in% c(2,5)) & merged_pvga$num_new_green %in% c(1,3)] <- "c: confirmed green oa"
merged_pvga$compliance_new_ta_2020[is.na(merged_pvga$compliance_new_ta_2020)] <- "No supported route"

merged_pvga$compliance_new_ta_2020 <- factor(merged_pvga$compliance_new_ta_2020, ordered = TRUE, levels = c("c: pure gold", "c: hybrid gold with a TA", "c: confirmed green oa", "No supported route"))

compliance_new_ta_2020_a <- merged_pvga %>%
  count(compliance_new_ta_2020, .drop = FALSE) %>%
  mutate(percent = round(n / sum(n) * 100,1), cml = round(cumsum(percent)))


  # Transformative agreements to end of 2021
merged_pvga$compliance_new_ta_2021 <- NA
merged_pvga$compliance_new_ta_2021[merged_pvga$num_fee %in% c(1,4)] <- "c: pure gold"
merged_pvga$compliance_new_ta_2021[merged_pvga$ta_split %in% c("2019","2020","2021") & merged_pvga$num_fee %in% c(2,5)] <- "c: hybrid gold with a TA"
merged_pvga$compliance_new_ta_2021[!merged_pvga$num_fee %in% c(1,4) & !(merged_pvga$ta_split %in% c("2019","2020","2021") & merged_pvga$num_fee %in% c(2,5)) & merged_pvga$num_new_green %in% c(1,3)] <- "c: confirmed green oa"
merged_pvga$compliance_new_ta_2021[is.na(merged_pvga$compliance_new_ta_2021)] <- "No supported route"

merged_pvga$compliance_new_ta_2021 <- factor(merged_pvga$compliance_new_ta_2021, ordered = TRUE, levels = c("c: pure gold", "c: hybrid gold with a TA", "c: confirmed green oa", "No supported route"))

compliance_new_ta_2021_a <- merged_pvga %>%
  count(compliance_new_ta_2021, .drop = FALSE) %>%
  mutate(percent = round(n / sum(n) * 100,1), cml = round(cumsum(percent)))


  # If Elsevier negotiated a TA

merged_pvga$compliance_new_ta_Elsevier <- NA
merged_pvga$compliance_new_ta_Elsevier[merged_pvga$num_fee %in% c(1,4)] <- "c: pure gold"
merged_pvga$compliance_new_ta_Elsevier[(merged_pvga$ta_split %in% c("2019","2020","2021") | merged_pvga$Publisher == "Elsevier") & merged_pvga$num_fee %in% c(2,5)] <- "c: hybrid gold with a TA"
merged_pvga$compliance_new_ta_Elsevier[!merged_pvga$num_fee %in% c(1,4) & !((merged_pvga$ta_split %in% c("2019","2020","2021") | merged_pvga$Publisher == "Elsevier") & merged_pvga$num_fee %in% c(2,5)) & merged_pvga$num_new_green %in% c(1,3)] <- "c: confirmed green oa"
merged_pvga$compliance_new_ta_Elsevier[is.na(merged_pvga$compliance_new_ta_Elsevier)] <- "No supported route"

merged_pvga$compliance_new_ta_Elsevier <- factor(merged_pvga$compliance_new_ta_Elsevier, ordered = TRUE, levels = c("c: pure gold", "c: hybrid gold with a TA", "c: confirmed green oa", "No supported route"))

compliance_new_ta_Elsevier_a <- merged_pvga %>%
  count(compliance_new_ta_Elsevier, .drop = FALSE) %>%
  mutate(percent = round(n / sum(n) * 100,1), cml = round(cumsum(percent)))


  # If TA coverage was 100%
merged_pvga$compliance_new_ta_100 <- NA
merged_pvga$compliance_new_ta_100[merged_pvga$num_fee %in% c(1,4)] <- "c: pure gold"
merged_pvga$compliance_new_ta_100[merged_pvga$num_fee %in% c(2,5)] <- "c: hybrid gold with a TA"
merged_pvga$compliance_new_ta_100[!merged_pvga$num_fee %in% c(1,4) & !merged_pvga$num_fee %in% c(2,4) & merged_pvga$num_new_green %in% c(1,3)] <- "c: confirmed green oa"
merged_pvga$compliance_new_ta_100[is.na(merged_pvga$compliance_new_ta_100)] <- "No supported route"

merged_pvga$compliance_new_ta_100 <- factor(merged_pvga$compliance_new_ta_100, ordered = TRUE, levels = c("c: pure gold", "c: hybrid gold with a TA", "c: confirmed green oa", "No supported route"))

compliance_new_ta_100_a <- merged_pvga %>%
  count(compliance_new_ta_100, .drop = FALSE) %>%
  mutate(percent = round(n / sum(n) * 100,1), cml = round(cumsum(percent)))

# Bind together articles
ta_impact_a <- bind_rows(compliance_new_ta_noTA_a, compliance_new_ta_2019_a, compliance_new_ta_2020_a, compliance_new_ta_2021_a, compliance_new_ta_Elsevier_a, compliance_new_ta_100_a) %>%
  mutate(ta_coverage = "No TAs") %>%
  unite(Compliance_Route, c('compliance_new_ta_noTA', 'compliance_new_ta_2019', 'compliance_new_ta_2020', 'compliance_new_ta_2021', 'compliance_new_ta_Elsevier', 'compliance_new_ta_100'), na.rm = TRUE) %>%
  relocate(5,1,2,3,4)

ta_impact_a[5:8,'ta_coverage'] <- "TAs by\n2019"
ta_impact_a[9:12,'ta_coverage'] <- "TAs by\n2020"
ta_impact_a[13:16,'ta_coverage'] <- "TAs by\n2021"
ta_impact_a[17:20,'ta_coverage'] <- "+ Elsevier\nTA"
ta_impact_a[21:24,'ta_coverage'] <- "Full TA\ncoverage"
ta_impact_a[c(1,5,9,13,17,21), 'Compliance_Route'] <- "Pure gold"
ta_impact_a[c(2,6,10,14,18,22), 'Compliance_Route'] <- "Hybrid gold with TA"
ta_impact_a[c(3,7,11,15,19,23), 'Compliance_Route'] <- "Confirmed green OA"
ta_impact_a[c(4,8,12,16,20,24), 'Compliance_Route'] <- "No confirmed supported route"

ta_impact_a$Compliance_Route <- factor(ta_impact_a$Compliance_Route, ordered = TRUE, levels = c("Pure gold", "Hybrid gold with TA", "Confirmed green OA", "No confirmed supported route"))
ta_impact_a$ta_coverage <- factor(ta_impact_a$ta_coverage, ordered = TRUE, levels = c("No TAs", "TAs by\n2019", "TAs by\n2020", "TAs by\n2021", "+ Elsevier\nTA", "Full TA\ncoverage"))

openxlsx::write.xlsx(as.data.frame(ta_impact_a), 'Output/Tables/ta_impact_a.xlsx')

# Create stacked bar chart for articles

ta_impact_a_bar <- ggplot(ta_impact_a, aes(fill=Compliance_Route, y=percent, x=ta_coverage)) +
  geom_bar(position = "stack", stat = "identity") +
  scale_fill_manual(values = c("#F08900", "#FBBB10", "#16978A", "#FF5A5A"), name = "Supported routes to\nOpen Access", labels = c("Gold OA in\nfully OA journal\n", "Gold OA in Hybrid\njournal with\nTransitional Agreement", "\nConfirmed green OA\n", "None identified under\ncurrent journal policies\n")) +
  theme(axis.title.x = element_text(face="bold", size=14, margin = margin(t=10)),
        axis.text.x  = element_text(vjust=0.5, size=14, angle = 90),
        axis.title.y = element_text(face="bold", size=14, margin = margin(t=10)),
        legend.text = element_text(size=14),
        legend.title = element_text(size=16, face="bold")) +
  labs(x="Introduction of new TAs", y= "% of articles")

ggsave(ta_impact_a_bar, filename = "Output/Charts/ta_impact_a_bar.png", width = 8, height = 5, dpi = 300)
